extraction/requirements.txt
requests
pandas
snowflake-connector-python
snowflake-sqlalchemy